# apps/backend/docs/swagger.yaml
basePath: /api
definitions:
...
  entity.UserWorkoutSession:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      duration_minutes:
        type: integer
      id:
        type: string
      location:
        type: string
      mood_rating:
        type: integer
      notes:
        type: string
      perceived_exertion_rating:
        type: integer
      plan_id:
        type: string
      scheduled_at:
        type: string
      session_name:
        type: string
      started_at:
        type: string
      status:
        $ref: '#/definitions/entity.WorkoutSessionStatus'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  entity.WorkoutSessionStatus:
    enum:
    - scheduled
    - in_progress
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - WorkoutSessionStatusScheduled
    - WorkoutSessionStatusInProgress
    - WorkoutSessionStatusCompleted
    - WorkoutSessionStatusCancelled
  v1.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Fitness application API for tracking nutrition, meals, and products
  title: Rebound API
  version: "1.0"
paths:
...
  /workout-sessions:
    post:
      consumes:
      - application/json
      description: Create a new workout session for a user
      parameters:
      - description: Workout session object
        in: body
        name: session
        required: true
        schema:
          $ref: '#/definitions/entity.UserWorkoutSession'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.UserWorkoutSession'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Create a new workout session
      tags:
      - workout-sessions
  /workout-sessions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a workout session by its ID
      parameters:
      - description: Workout session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Delete a workout session
      tags:
      - workout-sessions
    get:
      consumes:
      - application/json
      description: Get a workout session by its ID
      parameters:
      - description: Workout session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserWorkoutSession'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get a workout session by ID
      tags:
      - workout-sessions
    put:
      consumes:
      - application/json
      description: Update an existing workout session
      parameters:
      - description: Workout session ID
        in: path
        name: id
        required: true
        type: string
      - description: Workout session object
        in: body
        name: session
        required: true
        schema:
          $ref: '#/definitions/entity.UserWorkoutSession'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserWorkoutSession'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Update a workout session
      tags:
      - workout-sessions
  /workout-sessions/{id}/exercises:
    get:
      consumes:
      - application/json
      description: Get all exercises for a workout session
      parameters:
      - description: Workout session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.UserWorkoutSession'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get workout session exercises
      tags:
      - workout-sessions
    post:
      consumes:
      - application/json
      description: Add a new exercise to a workout session
      parameters:
      - description: Workout session ID
        in: path
        name: id
        required: true
        type: string
      - description: Exercise object
        in: body
        name: exercise
        required: true
        schema:
          $ref: '#/definitions/entity.UserWorkoutSession'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.UserWorkoutSession'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Add exercise to workout session
      tags:
      - workout-sessions
  /workout-sessions/exercises/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an exercise from a workout session
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Delete workout exercise
      tags:
      - workout-sessions
    put:
      consumes:
      - application/json
      description: Update an existing exercise in a workout session
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: string
      - description: Exercise object
        in: body
        name: exercise
        required: true
        schema:
          $ref: '#/definitions/entity.UserWorkoutSession'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserWorkoutSession'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Update workout exercise
      tags:
      - workout-sessions
  /workout-sessions/user/{userID}:
    get:
      consumes:
      - application/json
      description: Get all workout sessions for a specific user
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.UserWorkoutSession'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get user workout sessions
      tags:
      - workout-sessions
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"